Support for float prices

+++ comanda-peter/src/com/company/comanda/peter/server/GUIServiceImpl.java
@@ -89,7 +89,7 @@ GUIService {
                     item.getId().toString(),
                     item.getImageString(), 
                     item.getName(), 
-                    (new Integer(item.getPrice())).toString(),
+                    (new Float(item.getPrice())).toString(),
                     item.getDescription()});
         }
         return new PagedResult<String[]>(resultList, total);
